fix: Optimize averaging filter calculation
The averaging filter calculation in the image processing code has been optimized. The previous version used an extra variable to store the sum of pixel values, which was then divided by 9 to get the average. This has been simplified by directly calculating the average within the assignment statement, removing the need for an additional variable and making the code more efficient.